article-generator-hataraku/requirements.txt
anthropic>=0.40.0
openai>=1.99.0
python-dotenv>=1.0.0
requests>=2.32.0
pandas>=2.2.0
gspread>=6.0.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0
Pillow>=10.0.0
huggingface_hub>=0.25.0
